+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-beta30]
+- adds main cli: `clojure -M:dev -m scicloj.clay.v2.main`
+
 ## [2-beta29] - 2025-02-18
 - updated deps
 - prevented unnecessary `def`s in generated test namespaces